skbuild: Do not raise error if running clean command twice
This commit prevents this error from being reported if clean command
is executed twice:

```
running clean
removing '_skbuild/cmake-install'
error: [Errno 2] No such file or directory: '_skbuild/cmake-install'
```

+++ skbuild/command/clean.py
@@ -1,4 +1,6 @@
 
+import os
+
 try:
     from setuptools.command.clean import clean as _clean
 except ImportError:
@@ -18,6 +20,7 @@ class clean(set_build_base_mixin, new_style(_clean)):
         for dir_ in (cmaker.CMAKE_INSTALL_DIR,
                      cmaker.CMAKE_BUILD_DIR,
                      cmaker.SKBUILD_DIR):
-            log.info("removing '%s'", dir_)
-            if not self.dry_run:
+            if os.path.exists(dir_):
+                log.info("removing '%s'", dir_)
+            if not self.dry_run and os.path.exists(dir_):
                 rmtree(dir_)
